requirements.txt
Pillow
aiohttp
black
displayhatmini
dotenv